+++ .github/.testcoverage-local.yml
@@ -12,3 +12,4 @@ override:
 exclude:
   paths:
     - main\.go$
+    - main_config\.go$
