+++ pkg/testcoverage/logger/logger.go
@@ -2,17 +2,30 @@ package logger
 
 import (
 	"bytes"
+	"sync"
 
 	"github.com/rs/zerolog"
 )
 
 //nolint:gochecknoglobals // relax
 var (
-	Buffer bytes.Buffer
+	buffer bytes.Buffer
+	lock   sync.Mutex
 	L      zerolog.Logger
 )
 
-//nolint:gochecknoinits // relax
-func init() {
-	L = zerolog.New(&Buffer).With().Logger()
+func Init() { // coverage-ignore
+	L = zerolog.New(&buffer).With().Logger()
+}
+
+func Destruct() {
+	L = zerolog.Logger{}
+	buffer = bytes.Buffer{}
+}
+
+func Bytes() []byte {
+	lock.Lock()
+	defer lock.Unlock()
+
+	return buffer.Bytes()
 }
